Refactored CheepRepository to use CheepContext Co-authored-by: Mads <mvje@itu.dk>
