+++ src/Repositories/Usings.cs
@@ -1,1 +1,3 @@
 ﻿global using Chirp.Razor;
+global using Microsoft.EntityFrameworkCore;
+
